Annotate and fix platform-specific tests in System.IO.FileSystem.Watcher
A few tests for FileSystemWatcher are platform-specific, e.g. differences in event coalescing behavior causing a the event buffer to fill on Windows but not on Linux.  This commit annotates such tests as Windows-specific.  It also fixes up a few other tests that were mostly platform-independent but had a few pieces that were platform-dependent; the platform-dependent pieces are now conditionally invoked.