Expand test workaround for DST issue 1326 in Sys.IO.Compression
The System.IO.Comprssion tests verify the last modified metadata of
zip file entries. There was a workaround for an unknown DST issue that
only succeeded when the current locale supported daylight savings
time.  Expand the test workaround so it will work in all locales, and
update the comment in the test source to reference new github issue
https://github.com/dotnet/corefx/issues/1326