Fix anonymous memory maps on Unix to support multiple views
In our current implementation of anonymous (no mapName specified), non-persisted (no file provided) memory-mapped files, each time a view is created we call mmap with the anonymous option.  This effectively results in every view of the same MemoryMappedFile actually being distinct and not sharing any data.

This commit modifies the implementation so that anonymous, non-persisted memory-mapped files are instead backed by POSIX shared memory objects.  This makes it so that multiple views/streams over the same anonymous map correctly share data as expected.

I'd previously modified our tests to avoid exercising this unsupported path on Unix.  Now that it's supported, I've put the few tests back to the way they were so that we test multiple views over the same map correctly sharing data with each other.

In the future, we should be able to extend this to provide support for named, non-persisted maps.