Enable System.IO.Compression on Unix
This commit enables the System.IO.Compression library to work on Unix and fixes up the tests so that they all pass.

For now, the Unix compilation is using the managed deflater implementation.  We should subsequently fix this up to use zlib as does the Windows implementation, but since it already exists, using the managed implementation is a fine place to start.

A bunch of tests needed to be fixed up to handle differences between forward and backward slashes, as well as a couple of capitalization issues for case-sensitive OSes.