Initial commit of web encoders This is the first commit in a series to make web encoders part of core libraries. We have had some discussions with the ASP.NET team, and we concluded that the encoders are quite general purpose, and should be made part of the core.
This commit includes a subset of Microsoft.Framework.WebEncoders.Core that we deemed general purpose.
It excludes some APIs related to dependency injection and some ASP.NET specific helpers.

The only changes in the moved code are:
a) new namespace and assembly name (System.Text.Encodings.Web)
b) changes to project files required by the CoreFx repo build system
c) changes to embedded resource paths (as the directory structure has changed)

One caveat, some tests in the test project are disabled. They do pass, but the test harness logger
fails as it cannot output some exotic characters that are passed to InlineData (and so printed in the log).
The root cause of this is most probably in XmlEscape routine in
https://github.com/xunit/xunit/blob/40e153fe983977be8fe009384fae5c377bc557ff/src/common/XmlTestExecutionVisitor.cs.
I will fix this problem in a subsequent commit.