Override FlushAsync to return an already completed task
DeflateStream's Flush implementation is a nop, other than checking to make sure that the instance isn't disposed.  As such, not overriding FlushAsync will result in unnecessary work when the base Stream's FlushAsync queues a work item to call Flush.

This commit just adds a FlushAsync override that returns an already completed task.  There's a slight potential behavioral change here, in that now if FlushAsync is called on a disposed object, the FlushAsync call will synchronously throw the ObjectDisposedException, whereas previously the Task returned from FlushAsync would be faulted with ObjectDisposedException.  I think this is the better behavior, as the ODE highlights a misuse of the object; if this is problematic, though, we can remove or ammend this commit.