Move file system case sensitivity policy to common source
Move the file system case sensitivity policy from an internal
FileSystem property to the common linked-source internal class
PathInternals.  This allows System.IO.Compression.ZipFile to use
the same policy.  We will need to update this when we add the ability
to detect the case sensitivity of the file system, rather than
assuming based on operating system.