Update S.Diag.Contracts tests to verify behavior for DEBUG and !DEBUG
System.Diagnostics.Contracts is implemented in mscorlib.  In versions
of mscorlib compiled with the DEBUG property, the Contract.Assert
methods are active.  In versions of mscorlib compiled without the
DEBUG property, the Contract.Assert methods are no-ops.  This change
updates the test to verify the behavior based on if it was compiled
with or without the DEBUG property.  It's then important that the
configuration of the test (debug or retail) match that of the mscorlib
it is testing.