Add PlatformID.AnyUnix
There are ~110 tests that have been disabled with an ActiveIssue for PlatformID.Linux | PlatformID.OSX.  To simplify this and to avoid lots of additional churn when additional Unix-based platforms are added, I've added PlatformID.AnyUnix = Linux | OSX, and replaced all usage of Linux | OSX with just AnyUnix.  When new platforms are added, we can simply add to the enum, expand the meaning of AnyUnix, and most tests meant to be disabled for Unix will automatically be disabled.