Reimplement TaskExtensions.Unwrap
The currently checked in implementation of TaskExtensions.Unwrap isn't what's in .NET 4.5/6 desktop.  Rather, it's a slight improvement over what was shipped in .NET 4, before a bunch of optimizations were made to it for .NET 4.5.  In .NET 4.5, we effectively built the implementation into Task itself, taking advantage of Task internals to minimize allocations and other overheads; the Unwrap extension methods then had internals access and could delegate to that implementation in mscorlib.  For the separate implementation now in System.Threading.Tasks.dll in .NET Core, we don't have that internals access, and need to implement Unwrap on top of Task's public surface area.

Even so, we can do much better than the checked in implementation, and this commit overhauls it.  The current implementation incurs 17 allocations (!) per Unwrap, whereas mscorlib's implementation incurs only 1 (along with a commensurate reduction in the number of bytes allocated).  This implementation gets the number down from 17 to 4, again with a commensurate size reduction.  It also gets the single-threaded speed of the implementation (i.e. not taking into account the GC impact of the extra allocations) to effectively the same as the implementation in mscorlib, whereas the old implementation is measurably worse.

It also addresses an inconsistency between the implementations.  In .NET 4.5/6, when the inner task has a CancellationToken that causes the task to be canceled, that CancellationToken is stored into the proxy task, a fact visible due to that token propagating out of any OperationCanceledException thrown when await'ing the proxy task.  The current implementation doesn't have that, but rather the pre-.NET 4.5 behavior where the token wasn't marshaled.  This commit brings it back.