Fix light-up when running in a Windows Store App
Reflection invocation outside the set of approved Store App API results
in InvalidOperationException and not MemberAccessException.

Handle InvalidOperationException everywhwere we Invoke, GetValue, or
CreateDelegate.

There was also one place where we did so without any catch, and it can
be reasoned that we wouldn't ever get that far without having bailed
earlier. However, it is easier to reason about it by simply applying
the same pattern throughout and that's now the case.

Fixes #1591