Make EditAndContinueLogEntry fields in to properties
This had been flagged in prior API review but got skipped
when the EditAndContinue API was on the cutting block. It
came up again in the review of the unrelated EntityHandle
change.

Since we're now taking breaking changes to this same API,
take the opportunity to fix this in the name of consistency.