Fix System.Collections.Immutable.Tests on machines with VS 2015
The tests were failing because ResolveAssemblyReferences was not copying
System.Collections.Immutable from the project reference.  RAR was finding
the assembly in the GAC and skipping it.  That is never really safe to do for
CoreCLR so we should look into whether or not we can disable this globally.
For now I am using the same fix as https://github.com/nguerrera/corefx/commit/2bc73d7450fd7f301df28c18f975babf27ef3164.