Fix race condition in mutex test
One of the tests I added earlier today has a race condition that causes the test to fail.  The test relies on a task running on another thread, in order for the wait on the mutex to time out, but it's possible as written for the task to get inlined into the current thread when we wait on it, which will inadvertently cause the wait on the mutex to succeed.  A simple fix is just to use LongRunning, which will give the task a dedicated thread and ensure it's not inlined in a wait.