Update to latest CoreFx packages
Latest packages are now on dotnet-core myget feed.

This should address a many of the warnings in the build.  I will follow up with another change to address the rest.

This change brings in the latest build tools with updates to consume "latest" packages (* versions).

I've updated every project in the repo to use * versioning for dependencies.

The new packages can be used to resolve runtime assets as well (like clrcompression.dll) so I have removed the custom code in System.IO.Compression to copy this dll.  Instead the test project references the actual System.IO.Compression package and sets the correct runtime IDs to pull out the resources.  The System.IO.Compression.dll from the package is not used since the project reference takes precedence.

I had to workaround an issue in System.Xml.XmlSerializer with the current packages: System.Runtime.WindowsRuntime 4.0.0 has a depedendency on "windows" package which doesn't exist.  This was caused by our auto-generation task not filtering the winrt reference in the reference assembly's surface area.  I've fixed this in a separate change and we can remove the workaround once we have a new set of packages.  https://github.com/dotnet/corefx/issues/1761

I also had to add a dependency to System.Security.Cryptography.Hashing.Algorithms\tests due to an issue with DNX.  It's not honoring the dependency group with no targetFramework when another dependency group exists in the nuspec.  This is inconsistent with nuget v2 and should be fixed in DNX.  https://github.com/aspnet/dnx/issues/1881

[tfs-changeset: 1471885]