Preserve non-padding NULs in section-header names.
Section-header names are UTF-8 strings, and it is legal for UTF-8 strings
to contain NUL.  To comply with the PE/COFF spec, we must distinguish
between null-padding bytes and NUL-codepoint bytes.  Here we assume that
any NUL codepoint followed by a non-NUL codepoint is part of the name, and
not the start of padding.