More robust probing for MetadataToken API
In some contexts MetadataToken API may not be present at all or access to it could be restricted.
In particular since we use expression trees, the moment when we realize that the access is restricted could happen at the time of "Compile()" or at the time of the actual execution of a helper.

This change makes the strategy of switching to a fallback IsEquivalentTo helper more robust.

Regardless of what context we are in, user should not see binder throwing
"MemberInfo.get_MetadataToken() can't be used on this platform"

[tfs-changeset: 1477889]