Purpose: Bug 1180013 : [DataContractJsonSerializer]: Unable to De/Serialize 'ReadOnlyCollection<T>' in a Windows 10 Universal app
Cause: ReadOnlyCollection<T> used to be serializable throught [Serializable] attribute, which is no longer available in N and K.

Solution: Add ReadOnlyCollection<T> to DCS_KnownSerializableGenericTypes list so that DataContract can be created for ReadOnlyCollection<T>.

Note: This is for K only, the fix for N has been checked in. Unit tests have been checked in along with N fix.

[tfs-changeset: 1482134]