Test both managed/zlib System.IO.Compression implementations
System.IO.Compression has two implementations for compression/decompression: one based on zlib, and one implemented in managed code.  As we can use the managed implementation as a fallback in case zlib is unavailable, we need to ensure it's functioning properly.

This commit adds a debug-only static field to DeflateStream that supports overriding the normal logic for whether zlib or the managed implementation is used.  The DeflateStream and GZipStream tests are then augmented to force both implementations to be used.