Fix bug in Process' AsyncStreamReader.BeginReadLine
BeginReadLine is kicking off a Task to asynchronously run the ReadBuffer method.  ReadBuffer is itself an async method, returning a Task.  But Task.Factory.StartNew was being used to invoke ReadBuffer, such that StartNew was returning a Task<Task>, and StartNew doesn't automatically unwrap such inner Tasks.  This meant that the _readToBufferTask was completing as soon as the first await yielded, which meant that Process.WaitForExit wouldn't actually wait for the async reading to complete.  This change simply switches to using Task.Run, which automatically unwraps the inner task due to its overload that takes a Func<Task> delegate.