Modify ZipFile tests to remove temporary garbage
The System.IO.Compression.ZipFile tests were leaving behind lots of temporary files and directories.  I've modified the tests to delete these when the tests completes.

As long as I was modifying the tests, I also boosted the code coverage of ZipFile and ZipFileExtensions to 100% (the library itself isn't at 100% because of the SR file and some unused code in a shared common file).

And I fixed up a few "String"=>"string"s I saw along the way.