Separate BCryptGenRandom out of Cng.cs
As part of #1739, this commit begins to refactor Cng.cs, starting by pulling out the functionality needed for RandomNumberGenerator.  The assembly was being filled with all of the Cng code, when in reality it only needed a very small amount of it.  And since the assembly doesn't have any string resources, I also modified the .csproj to suppress the buildtools inclusion of the common resources support.

As a result of the removals, this change boosts the code coverage number of System.Security.Cryptography.RandomNumberGenerator.dll from 13% to 96%.

(Note that there are still copies of the code in Interop.NTSTATUS.cs in Cng.cs, due to other dependencies in Cng.cs that I didn't want to change.  As we refactor Cng.cs further, we can remove those copies.)