Fixes bug 1958
Another remaining test for sameness that we agreed to remove (i.e. cost to synchronize too high
in comparison to the benefit)

There are a few other tests like that (they all test sameness of codepointfilter creation).
I will review them as a separate workitem as they are probably also not very reliable. But for now,
I am removing this test as it is actually failing during runs.