Disabling System.Diagnostics.Process.MainModule test against Issue#1896
EnumProcessModules which is used internally for getting the list of process modules
and main module. This API is a best effort API which may not work in certain scenarios.
I am still debating whether we should have test for this in our CI.
For now, I am disabling this against an Issue#1896 while I gather more data
on whether it makes sense to remove the test completely.