Optimize ImmutableArray<T>.RemoveRange for 0/1 arguments
RemoveRange(IEnumerable<T>, ...) is fairly expensive allocation-wise, allocating a SortedSet<int>, an enumerator, a new resulting array, etcr.  We can easily avoid such costs if the immutable items array contains 0 or 1 elements.

RemoveRange(int, int) also allocates a new array.  We can avoid that cost if the length of the requested range to remove is empty.