First round of changes based on API review feedback
We did an API review of the APIs. The review feedback is captured at
https://github.com/dotnet/apireviews/tree/master/2015-06-16-web-encoders

This set of changes corresponds to feedback that was both clear (not "think about ...")
and did not require moving files around (e.g. namespace name changes).
Those will possibly be implemented in a subsequent commit.

The main changes in this commit are:
- Merge in extension methods into TextEncoder
- Make default encoders internal
- CodeFilter mutators should be void
- Remove ICodePointFilter
- Removed string overloads from CodePointFilter