Remove SignAssembly=false from test assemblies
Setting SignAssembly to false breaks loading our test assemblies
in XUnit using the current folder structure we use.  It doesn't show
up in the open, because we always sign the assemblies, ignoring
this setting. But in our official drops, we respect the property, and
then are unable to run these 12 tests.