Properly reset HashAlgorithm objects using OpenSSL.
The keyed hash algorithms need to use HMAC_Init_ex, which supports an in-place reset.  The unkeyed algorithms have to be reinitialized completely using EVP_DigestInit_ex, passing NULL leads to a SIGSEGV, so persist the algorithm pointer.

Added reusability tests, and enabled the DeriveBytes tests that were disabled due to this issue.