Fix a few issues with AssertWithCallerAttributes
This file exists to help debug xunit Asserts on Unix, providing filenames and line numbers for failed assertions.  It's meant to be usable just by including the file in the test project, but a couple of the signatures with optional parameters caused problems when I tried to do that in System.Runtime.Extensions.Tests.  This provides a few minor tweaks to address that.