Add checks for infinity to Complex.Pow verification
Several times now the standardNumericalFunctions_PowTest.RunTests_RandomValidValues test has failed in CI.  The issue is that random values are generated that result in so large or small values that the Complex.Pow(x,y) computation results in infinity.  The test then tries to verify the result by comparing it to the result of Complex.Exp(y*Complex.Log(x)).  It needs to do an approximate comparison, but that approximation doesn't take infinity into account.  This commit attempts to address that by checking to see if one of the values is infinity, and if it is, rather than subtracting one value from the other (which will result in infinity), it tries to nudge the other value a bit to see if it becomes a matching infinity.