Convert UTF-16 encoded files into UTF8
UTF16 encoded files trip up both Git and GitHub since the huristic they
use to determine if a file is binary or not depends on the existence of
null bytes in a file, which is common in UTF-16 encoded source.

This converts all the UTF-16 encoded files into UTF-8 with an explicit
byte order marker.