Fix Process.WaitForExit premature timeout on Unix
The Process.WaitForExit implementation on Unix is mixing up units, resulting in WaitForExits with timeouts exiting much earlier than the specified timeout.  The code is using Stopwatch.GetTimestamp to get a current timestamp, subtracting the current timestamp from one gathered when the wait started, and then subtracting that difference from the specified timeout.  But the specified timeout is in milliseconds, and the timestamp is in ticks.  The fix just converts the difference from ticks to milliseconds before subtracing from the timeout value.